requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
yfinance==0.2.18
plotly==5.17.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
requests==2.31.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
alpaca-trade-api==3.1.1
websocket-client==1.6.4
python-dotenv==1.0.0
ta==0.10.2
scipy==1.11.4
scikit-learn==1.7.0
